Move stateful components into container directory
