+++ src/components/__tests__/Categories.test.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { cleanup, screen } from '@testing-library/react';
 import { render } from '../../test-utils';
 import '@testing-library/jest-dom/extend-expect';
-import Categories from '../Categories';
+import Categories from '../../container/Categories';
 
 afterEach(cleanup);
