+++ src/container/CategoryMeals.js
@@ -4,8 +4,8 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { Card, Text, Image } from '@mantine/core';
 import { Link } from '@reach/router';
-import Loading from './Loading';
-import NotFoundPage from './NotFound';
+import Loading from '../components/Loading';
+import NotFoundPage from '../components/NotFound';
 import { getMealsByCategory } from '../selectors';
 import { loadMeals, filter } from '../actions/index';
 import './CategoryMeals.css';
