+++ src/index.js
@@ -4,7 +4,7 @@ import { createStore, applyMiddleware, compose } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import './index.css';
-import App from './components/App';
+import App from './container/App';
 import reducer from './reducer';
 import { loadCategories, loadMealDetails } from './actions';
